Add '-p 12' option to SCRM.
This increases the precision for the position of segregating sites.  With the
default of 6, I cannot calculate the site position with base pair precision
when simulating a 100mb chromosome.  I could use the '-SC abs' option to
output the position in base pairs, but then I'd need to change downstream
scripts.