Add script to create ramdisk.
All this simulated data and fragmentation for MSMC creates an enormous
number of small files.  On an HDD, this is exceptionally inefficient.
On a ramdisk, this is very fast.  The zram block device is compressed, making
this reasonable for the 20-30gb of data, which is 15% of the size
when compressed.